refactoring extractNtf
