+++ src/scripts/model/Notification.js
@@ -128,4 +128,24 @@ export default class Notification {
       `${key}=${searchObject[key]}`
     ).join('&');
   }
+
+  static fromJson(jsonResponse) {
+    /*jshint -W069 */
+    if (jsonResponse['success']) {
+      let notifications = [];
+      for (let key in Notification.MODULE) {
+        if (jsonResponse[key].length !== 0) {
+          let moduleId = Notification.MODULE[key];
+          jsonResponse[key].forEach(rawNotification => {
+            let notification = new Notification(moduleId, rawNotification.id, rawNotification.url);
+            notifications.push(notification);
+          });
+        }
+      }
+      return notifications;
+    } else {
+      console.log('something went wrong but what ? session time out maybe ?', jsonResponse);
+      throw 'extract failed';
+    }
+  }
 }
